#540 Move lastModified checks to new changes_hunter, fix bug with always recompiling all pseudopatterns
